requirements.txt
Django~=3.2.5
pip~=20.2.3
pytz~=2021.1
sqlparse~=0.4.1
psycopg2~=2.9.1
asgiref~=3.4.1
gunicorn~=20.1.0
setuptools~=49.2.1
djangorestframework~=3.12.4
PyJWT~=1.7.1
djangorestframework-jwt~=1.11.0